% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeptideMHC2.R
\name{bindingVector}
\alias{bindingVector}
\title{Binding vector}
\usage{
bindingVector(peptide, MHC, locus)
}
\arguments{
\item{peptide}{The peptide in terms of amino acids sequence}

\item{MHC}{The beta sequence of the mhc-ii molecule}

\item{locue}{The type of mhc-ii molecule}
}
\value{
A dataframe contains both the first- and second-oreder
 Hamiltonians
}
\description{
This fucntion processes the  binding vector between the peptide and mhc-ii
molecule from all possible registers on the giving peptide.
}
\examples{
This is how to call this function
binding Vector('KLAMDVNNKHFRT', 'DRB1_0101', 'DRB')

}
