% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeptideMHC2.R
\name{getInputVector}
\alias{getInputVector}
\title{Getting the actual binding vector}
\usage{
getInputVector(peptide, mhcName, locus)
}
\arguments{
\item{peptide}{The peptide sequences}

\item{mhcName}{The the name of the mhc-ii molecule}

\item{locue}{The type of mhc-ii molecule}
}
\value{
A dataframe of two columns, the first column contains the names of
 the Hamiltonians and the second column contains their obtained values.
}
\description{
This function creates the actual binding vector for peptide:MHC-II
interaction.
}
\examples{
 getInputVector('KHVNWQERTPLK', 'DRB1_0404_seq', 'DRB')

}
