% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeptideMHC2.R
\name{parametersVector}
\alias{parametersVector}
\title{Parameters (Hamiltonians) vector}
\usage{
parametersVector(locus)
}
\arguments{
\item{locus}{One of the three loci: DQ, DP, or DRB.}
}
\value{
A dataframe of two columns, the first column contains the names of
 the Hamiltonians and the second column contains their values. By default
 these values are set to zero.
}
\description{
This function setups a vector containing Hamiltonians
that are corresponding to the given locus.
}
\examples{
This is how to call this function
  parametersVector('DQ')

}
