% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeptideMHC2.R
\name{predictBinding}
\alias{predictBinding}
\title{Predicts peptide:MHC-II binding interaction.}
\usage{
predictBinding(pep, mhc_name, locus, opt = 1)
}
\arguments{
\item{pep}{The amino acids sequence of the peptide of length =>9 nano mer.}

\item{mhc_name}{The name of mhc-ii molecule.}

\item{locus}{One of the three standard loci: DQ, DP, and DRB}

\item{opt}{This parameter determines the prediction type (intra-allele
or trans-alleles). 1 stands for intra-alleleic and 2 for trans-allelic.
The default value of this is 1.}
}
\value{
A vector cantaining four items: the peptide, the mhc-ii molecule
name, the probability of the binding, and the binary value (0 or 1),
 respectively.
}
\description{
This function predicts the outcome of interaction between peptide
and mhc-ii
molecule by calculating the energy of the binding from all possible
configurations between the peptide and the mhc-ii ligand. It derives such
configurations by using function getInputVector
}
\examples{
predictBinding('AIGIITLYLGAVVQA', 'DRB1_1501',  'DRB', 1 )
predictBinding('ITKLGAKPDGKTDCT', 'DQB10201',  'DQ', 2)
}
