% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeptideMHC2.R
\name{predictBindingSet}
\alias{predictBindingSet}
\title{Make predictions for set of peptides and mhc-ii molecules}
\usage{
predictBindingSet(dataset, opt = 1)
}
\arguments{
\item{dataset}{Dataframe of three columns; containing the peptide, name
of the mhc-ii molecule, and the name of locus, respectively. The order of
columns must be maintained.}

\item{opt}{Determines the prediction type (intra-allele or trans-alleles).
1 stands for intra-alleleic and 2 for trans-allelic. The default value
of this is 1.}
}
\value{
A dataframe contains four columns: the peptide, the mhc-ii molecule
name, the probability of the binding, and the binary value (0 or 1)
respectively.
}
\description{
This function predicts binding affinities for a dataframe  containing set
of peptides with set of the targeted mhcii Molcules, associated with their
perspective loci. It generalization of predictionBinding function.
}
\examples{
 predictBindingSet <- function(dataset, opt =2)

}
