% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeptideMHC2.R
\name{secondOrderHamiltonians}
\alias{secondOrderHamiltonians}
\title{Second order Hamiltonians}
\usage{
secondOrderHamiltonians(peptide, MHC, locus)
}
\arguments{
\item{peptide}{The peptide sequence}

\item{MHC}{The beta sequence of the targeted mhc-ii molecule}

\item{locus}{The intended mhc-ii locus, and it must be on of the
following loci: DQ, DP, and  DRB.}
}
\value{
A vector conatining the  pairwise interacting Hamiltonians
}
\description{
This function  processes the second-order Hamiltonians between the peptide
and mhc-ii molecule.
}
\examples{
This is how to use this function
 secondOrderHamiltonians(pep, mhc, drb)

}
