% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeptideMHC2.R
\name{sliding}
\alias{sliding}
\title{Getting all possible frames}
\usage{
sliding(peptide)
}
\arguments{
\item{peptide}{The amino acids sequence of  peptide of any length >= 9 mer.}
}
\value{
A vector containing all the possible frames on the given peptide.
}
\description{
This fucntion extracts all the possible frames  from the
 giveing peptide (frames of 9 mers).
}
\examples{
This is how to call this function
    sliding('HKLMNARTIPPL')

}
