% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeptideMHC2.R
\name{supportedMolecules}
\alias{supportedMolecules}
\title{Supported MHC class II molecules}
\usage{
supportedMolecules()
}
\value{
A dataframe containing two columns, the first column shows MHC-II
molecule name and the second column shows its alias neme. Please use
the names that appear in first column
}
\description{
This function shows  MHC-II molecules  covered within this package.
}
\examples{
## What are the MHC-II molecules covered by this package?
supportedMolecules()

}
